Split the remaining refactorings, except EncapsulateField
