+++ Rubberduck.Refactorings/Abstract/BaseRefactoringBase.cs
