+++ Rubberduck.Refactorings/Abstract/IBaseRefactoring.cs
@@ -1,4 +1,6 @@
-﻿namespace Rubberduck.Refactorings
+﻿using Rubberduck.Parsing.Rewriter;
+
+namespace Rubberduck.Refactorings
 {
     /// <summary>
     /// The heart of a refactoring: this part of the refactoring performs the actual transformation of the code once all necessary information has been gathered.
@@ -13,7 +15,7 @@
         void Refactor(TModel model);
     }
 
-    public interface IBaseRefactoringWithPreview<in TModel> : IBaseRefactoring<TModel>
+    public interface IRefactoringPreviewProvider<in TModel>
         where TModel : class, IRefactoringModel
     {
         /// <summary>
@@ -22,5 +24,16 @@
         /// <param name="model">The model used by the refactoring containing all information needed to specify what to do.</param>
         /// <returns>Preview of the refactored code</returns>
         string Preview(TModel model);
+    }
+
+    public interface ICodeOnlyBaseRefactoring<in TModel> : IBaseRefactoring<TModel>
+        where TModel : class, IRefactoringModel
+    {
+        /// <summary>
+        /// Performs the refactoring according to the model and using the provided rewrite session.
+        /// </summary>
+        /// <param name="model">The model specifying all parameters of the refactoring</param>
+        /// <param name="rewriteSession">Rewrite session used to manipulate the code (Does not get executed.)</param>
+        void Refactor(TModel model, IRewriteSession rewriteSession);
     }
 }
