+++ Rubberduck.Refactorings/Abstract/IRefactoring.cs
