+++ Rubberduck.Refactorings/Abstract/IRefactoringDialog.cs
