+++ Rubberduck.Refactorings/Abstract/IRefactoringModel.cs
