+++ Rubberduck.Refactorings/Abstract/IRefactoringPresenter.cs
