+++ Rubberduck.Refactorings/Abstract/IRefactoringPresenterFactory.cs
