+++ Rubberduck.Refactorings/Abstract/IRefactoringView.cs
