+++ Rubberduck.Refactorings/Abstract/IRefactoringViewModel.cs
