+++ Rubberduck.Refactorings/Abstract/InteractiveRefactoringBase.cs
