+++ Rubberduck.Refactorings/Abstract/RefactoringBase.cs
