+++ Rubberduck.Refactorings/ImplementInterface/AddInterfaceImplementations/AddInterFaceImplementationsBaseRefactoring.cs
@@ -1,0 +1,98 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Rubberduck.Parsing.Grammar;
+using Rubberduck.Parsing.Rewriter;
+using Rubberduck.Parsing.Symbols;
+using Rubberduck.Refactorings.ImplementInterface;
+
+namespace Rubberduck.Refactorings.AddInterfaceImplementations
+{
+    public class AddInterFaceImplementationsBaseRefactoring : CodeOnlyBaseRefactoringBase<AddInterfaceImplementationsModel>
+    {
+        private const string MemberBody = "    Err.Raise 5 'TODO implement interface member";
+
+        public AddInterFaceImplementationsBaseRefactoring(IRewritingManager rewritingManager) 
+            : base(rewritingManager)
+        {}
+
+        public override void Refactor(AddInterfaceImplementationsModel model, IRewriteSession rewriteSession)
+        {
+            var rewriter = rewriteSession.CheckOutModuleRewriter(model.TargetModule);
+            AddItems(model.Members, rewriter, model.InterfaceName);
+        }
+
+        private void AddItems(IEnumerable<Declaration> missingMembers, IModuleRewriter rewriter, string interfaceName)
+        {
+            var missingMembersText = missingMembers
+                .Aggregate(string.Empty, (current, member) => current + Environment.NewLine + GetInterfaceMember(member, interfaceName));
+
+            rewriter.InsertAfter(rewriter.TokenStream.Size, Environment.NewLine + missingMembersText);
+        }
+
+        private string GetInterfaceMember(Declaration member, string interfaceName)
+        {
+            var template = string.Join(Environment.NewLine, Tokens.Private + " {0}{1} {2}{3}", MemberBody, Tokens.End + " {0}", string.Empty);
+            var signature = $"{interfaceName}_{member.IdentifierName}({string.Join(", ", GetParameters(member))})";
+            var asType = $" {Tokens.As} {member.AsTypeName}";
+
+            switch (member.DeclarationType)
+            {
+                case DeclarationType.Procedure:
+                    return string.Format(template, Tokens.Sub, string.Empty, signature, string.Empty);
+                case DeclarationType.Function:
+                    return string.Format(template, Tokens.Function, string.Empty, signature, asType);
+                case DeclarationType.PropertyGet:
+                    return string.Format(template, Tokens.Property, $" {Tokens.Get}", signature, asType);
+                case DeclarationType.PropertyLet:
+                    return string.Format(template, Tokens.Property, $" {Tokens.Let}", signature, string.Empty);
+                case DeclarationType.PropertySet:
+                    return string.Format(template, Tokens.Property, $" {Tokens.Set}", signature, string.Empty);
+                case DeclarationType.Variable:
+                    var members = new List<string>
+                    {
+                        string.Format(template, Tokens.Property, $" {Tokens.Get}", $"{interfaceName}_{member.IdentifierName}()", asType)
+                    };
+
+                    if (member.AsTypeName.Equals(Tokens.Variant) || !member.IsObject)
+                    {
+                        members.Add(string.Format(template, Tokens.Property, $" {Tokens.Let}", signature, string.Empty));
+                    }
+
+                    if (member.AsTypeName.Equals(Tokens.Variant) || member.IsObject)
+                    {
+                        members.Add(string.Format(template, Tokens.Property, $" {Tokens.Set}", signature, string.Empty));
+                    }
+
+                    return string.Join(Environment.NewLine, members);
+            }
+
+            return string.Empty;
+        }
+
+        private IEnumerable<Parameter> GetParameters(Declaration member)
+        {
+            if (member.DeclarationType == DeclarationType.Variable)
+            {
+                return new List<Parameter>
+                {
+                    new Parameter
+                    {
+                        Accessibility = Tokens.ByVal,
+                        Name = "rhs",
+                        AsTypeName = member.AsTypeName
+                    }
+                };
+            }
+
+            return ((ModuleBodyElementDeclaration)member).Parameters.Select(p => new Parameter
+            {
+                Accessibility = ((VBAParser.ArgContext)p.Context).BYVAL() != null
+                    ? Tokens.ByVal
+                    : Tokens.ByRef,
+                Name = p.IdentifierName,
+                AsTypeName = p.AsTypeName
+            });
+        }
+    }
+}
