+++ Rubberduck.Refactorings/ImplementInterface/ImplementInterfaceModel.cs
@@ -1,0 +1,16 @@
+﻿using Rubberduck.Parsing.Symbols;
+
+namespace Rubberduck.Refactorings.ImplementInterface
+{
+    public class ImplementInterfaceModel : IRefactoringModel
+    {
+        public ClassModuleDeclaration TargetInterface { get; }
+        public ClassModuleDeclaration TargetClass { get; }
+
+        public ImplementInterfaceModel(ClassModuleDeclaration targetInterface, ClassModuleDeclaration targetClass)
+        {
+            TargetInterface = targetInterface;
+            TargetClass = targetClass;
+        }
+    }
+}
