+++ Rubberduck.Refactorings/IntroduceField/IntroduceFieldModel.cs
@@ -1,0 +1,14 @@
+﻿using Rubberduck.Parsing.Symbols;
+
+namespace Rubberduck.Refactorings.IntroduceField
+{
+    public class IntroduceFieldModel : IRefactoringModel
+    {
+        public Declaration Target { get; }
+
+        public IntroduceFieldModel(Declaration target)
+        {
+            Target = target;
+        }
+    }
+}
