+++ Rubberduck.Refactorings/MoveCloserToUsage/MoveCloserToUsageModel.cs
@@ -1,0 +1,14 @@
+﻿using Rubberduck.Parsing.Symbols;
+
+namespace Rubberduck.Refactorings.MoveCloserToUsage
+{
+    public class MoveCloserToUsageModel : IRefactoringModel
+    {
+        public Declaration Target { get; }
+
+        public MoveCloserToUsageModel(Declaration target)
+        {
+            Target = target;
+        }
+    }
+}
