+++ RubberduckTests/Commands/RefactorCommands/ImplementInterfaceCommandTests.cs
@@ -3,6 +3,7 @@ using NUnit.Framework;
 using Rubberduck.Interaction;
 using Rubberduck.Parsing.Rewriter;
 using Rubberduck.Parsing.VBA;
+using Rubberduck.Refactorings.AddInterfaceImplementations;
 using Rubberduck.Refactorings.ImplementInterface;
 using Rubberduck.UI.Command;
 using Rubberduck.UI.Command.Refactorings;
@@ -56,7 +57,9 @@ Dim b As Variant";
         protected override CommandBase TestCommand(IVBE vbe, RubberduckParserState state, IRewritingManager rewritingManager, ISelectionService selectionService)
         {
             var msgBox = new Mock<IMessageBox>().Object;
-            var refactoring = new ImplementInterfaceRefactoring(state, rewritingManager, selectionService);
+            var addImplementationsBaseRefactoring = new AddInterFaceImplementationsBaseRefactoring(rewritingManager);
+            var baseRefactoring = new ImplementInterfaceBaseRefactoring(addImplementationsBaseRefactoring, rewritingManager);
+            var refactoring = new ImplementInterfaceRefactoring(baseRefactoring, state, rewritingManager, selectionService);
             var notifier = new ImplementInterfaceFailedNotifier(msgBox);
             return new RefactorImplementInterfaceCommand(refactoring, notifier, state, selectionService);
         }
