+++ RubberduckTests/Commands/RefactorCommands/IntroduceParameterCommandTests.cs
@@ -44,7 +44,8 @@ End Property";
         {
             var msgBox = new Mock<IMessageBox>().Object;
             var selectedDeclarationProvider = new SelectedDeclarationProvider(selectionService, state);
-            var refactoring = new IntroduceParameterRefactoring(state, msgBox, rewritingManager, selectionService, selectedDeclarationProvider);
+            var baseRefactoring = new IntroduceParameterBaseRefactoring(state, rewritingManager);
+            var refactoring = new IntroduceParameterRefactoring(baseRefactoring, msgBox, rewritingManager, selectionService, selectedDeclarationProvider);
             var notifier = new IntroduceParameterFailedNotifier(msgBox);
             return new RefactorIntroduceParameterCommand(refactoring, notifier, state, selectionService, selectedDeclarationProvider);
         }
