+++ RubberduckTests/Commands/RefactorCommands/RemoveParametersCommandTests.cs
@@ -170,7 +170,8 @@ End Property";
             uiDispatcherMock
                 .Setup(m => m.Invoke(It.IsAny<Action>()))
                 .Callback((Action action) => action.Invoke());
-            var refactoring = new RemoveParametersRefactoring(state, factory, rewritingManager, selectionService, selectedDeclarationProvider, uiDispatcherMock.Object);
+            var baseRefactoring = new RemoveParameterBaseRefactoring(state, rewritingManager);
+            var refactoring = new RemoveParametersRefactoring(baseRefactoring, state, factory, rewritingManager, selectionService, selectedDeclarationProvider, uiDispatcherMock.Object);
             var notifier = new RemoveParameterFailedNotifier(msgBox);
             return new RefactorRemoveParametersCommand(refactoring, notifier, state, selectionService, selectedDeclarationProvider);
         }
